+++ bld/hlpview/c/search.c
@@ -264,7 +264,7 @@ unsigned long HelpFindTopicOffset( HelpHdl hdl, char *topic )
 HelpHdl InitHelpSearch( HelpFp fp )
 {
     HelpHdl     hdl;
-    size_t      len;
+    unsigned    len;
     char        *topic;
     char        *description;
     uint_16     str_cnt;
